refactor: extract project report and repo routes
